+++ app/pages/hardwaretickets/index.vue
@@ -13,6 +13,12 @@
                             @change="hardwaretickets.search()">
                         </FormInput>
                     </div>
+                    <div class="col">
+                        <FormInput type="select" :options="priorities" name="priorityFilter"
+                            v-model="hardwaretickets.query.priority" placeholder="All Priorities"
+                            @change="hardwaretickets.search()">
+                        </FormInput>
+                    </div>
                     <div>
                         <FormInput type="select" :options="allhardwareitems.recordsAsOptions('props', 'name')"
                             name="hardwareItemIdFilter" placeholder="All Equipment"
@@ -49,6 +55,7 @@
                             </NuxtLink>
                         </template>
                     </Column>
+                    <Column field="props.priority" header="Priority" sortable />
                     <Column field="props.type" header="Type" sortable>
                         <template #body="slotProps">
                             {{ enum2label(slotProps.data.props.type, hardwaretickets.propsOptions.type) }}
@@ -124,10 +131,17 @@
 definePageMeta({
     middleware: ['auth', 'maintenance-production-manager', 'all-workaround']
 });
+const priorities = ref([
+    { value: 1, label: '1' },
+    { value: 2, label: '2' },
+    { value: 3, label: '3' },
+    { value: 4, label: '4' },
+    { value: 5, label: '5' },
+]);
 const hardwaretickets = useVingKind({
     listApi: `/api/${useRestVersion()}/hardwaretickets`,
     createApi: `/api/${useRestVersion()}/hardwaretickets`,
-    query: { includeMeta: true, sortBy: 'createdAt', sortOrder: 'desc', hardwareTaskId: '', hardwareItemId: '', type: '', status: 'unresolved', description: '' },
+    query: { includeMeta: true, sortBy: 'createdAt', sortOrder: 'desc', hardwareTaskId: '', hardwareItemId: '', type: '', priority: '', status: 'unresolved', description: '' },
     newDefaults: { description: '', type: 'needs_help', severity: 'working', status: 'unresolved', submittedBy: '', hardwareTaskId: '', hardwareItemId: '', id: '' },
     onCreate: (data) => { return navigateTo(`/hardwaretickets/${data.props.id}`) },
 });